homework/week07_08/transaction-stream/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
gqlalchemy==1.0.6
    # via -r requirements.in
kafka-python==2.0.2
    # via -r requirements.in
networkx==2.6.3
    # via gqlalchemy
pymgclient==1.1.0
    # via gqlalchemy
